Move HasHome out of Binder
